services/orchestrator/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
httpx==0.25.2
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4
python-multipart==0.0.6
aiofiles==23.2.1
redis==5.0.1
celery==5.3.4
